feat: add class keyboard, change style
