+++ src/index.js
@@ -1,14 +1,6 @@
 import KEYS from './data/keys';
-import Key from './js/key';
+import Keyboard from './js/keyboard';
 
-const keyBoard = document.createElement('div');
-keyBoard.classList.add('key-board');
-const { body } = document;
-body.append(keyBoard);
-
-KEYS.forEach((element) => {
-  const key = new Key(element).render();
-  keyBoard.append(key);
-});
+window.addEventListener('load', () => new Keyboard(KEYS).addKeyBoard());
 
 document.addEventListener('keydown', (e) => console.log(e));
