+++ src/js/key.js
@@ -1,12 +1,19 @@
 export default class Key {
-  constructor(key) {
+  constructor(key, language) {
     this.key = key;
-    this.language = 'en';
+    this.language = language;
+  }
+
+  addDataAtribute(element) {
+    Object.entries(this.key).forEach(([key, value]) => {
+      element.setAttribute(`data-${key}`, value);
+    });
   }
 
   render() {
     const key = document.createElement('div');
     key.classList.add('key');
+    this.addDataAtribute(key);
     const text = document.createElement('span');
     text.innerText = this.key[`key-${this.language}`];
     text.classList.add('key__text');
